RadarSDK: RRM 1.0.0

+++ oal/libs/kernel/common/src/linux/os_oal_atomic.c
@@ -1,11 +1,11 @@
 /*
- * Copyright 2018 NXP
+ * Copyright 2018, 2021 NXP
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-#include <oal_atomic.h>
 #include <oal_log.h>
+#include <oal_atomic.h>
 
 int32_t OAL_SetAtomic(struct OAL_Atomic *apAtomic, uint32_t aNewValue,
                       uint32_t *apOldValue)
@@ -15,10 +15,10 @@ int32_t OAL_SetAtomic(struct OAL_Atomic *apAtomic, uint32_t aNewValue,
 	OAL_CHECK_NULL_PARAM(apAtomic, set_atomic_exit);
 
 	if (apOldValue != NULL) {
-		*apOldValue = atomic_read(&apAtomic->mOSAtomic);
+		*apOldValue = (uint32_t)atomic_read(&apAtomic->mOSAtomic);
 	}
 
-	atomic_set(&apAtomic->mOSAtomic, aNewValue);
+	atomic_set(&apAtomic->mOSAtomic, (int32_t)aNewValue);
 set_atomic_exit:
 	return lRet;
 }
@@ -30,7 +30,7 @@ int32_t OAL_GetAtomic(struct OAL_Atomic *apAtomic, uint32_t *apValue)
 	OAL_CHECK_NULL_PARAM(apAtomic, get_atomic_exit);
 	OAL_CHECK_NULL_PARAM(apValue, get_atomic_exit);
 
-	*apValue = atomic_read(&apAtomic->mOSAtomic);
+	*apValue = (uint32_t)atomic_read(&apAtomic->mOSAtomic);
 get_atomic_exit:
 	return lRet;
 }
@@ -42,7 +42,7 @@ int32_t OAL_IncrementAtomic(struct OAL_Atomic *apAtomic, uint32_t *apOldValue)
 	OAL_CHECK_NULL_PARAM(apAtomic, inc_atomic_exit);
 
 	if (apOldValue != NULL) {
-		*apOldValue = atomic_read(&apAtomic->mOSAtomic);
+		*apOldValue = (uint32_t)atomic_read(&apAtomic->mOSAtomic);
 	}
 
 	atomic_inc(&apAtomic->mOSAtomic);
@@ -57,7 +57,7 @@ int32_t OAL_DecrementAtomic(struct OAL_Atomic *apAtomic, uint32_t *apOldValue)
 	OAL_CHECK_NULL_PARAM(apAtomic, dec_atomic_exit);
 
 	if (apOldValue != NULL) {
-		*apOldValue = atomic_read(&apAtomic->mOSAtomic);
+		*apOldValue = (uint32_t)atomic_read(&apAtomic->mOSAtomic);
 	}
 
 	atomic_dec(&apAtomic->mOSAtomic);
@@ -74,7 +74,7 @@ int32_t OAL_AddAtomic(struct OAL_Atomic *apAtomic, uint32_t aAddedValue,
 	OAL_CHECK_NULL_PARAM(apAtomic, add_atomic_exit);
 
 	if (apOldValue != NULL) {
-		*apOldValue = atomic_read(&apAtomic->mOSAtomic);
+		*apOldValue = (uint32_t)atomic_read(&apAtomic->mOSAtomic);
 	}
 
 	atomic_add((int32_t)aAddedValue, &apAtomic->mOSAtomic);
@@ -91,7 +91,7 @@ int32_t OAL_SubtractAtomic(struct OAL_Atomic *apAtomic, uint32_t aSubValue,
 	OAL_CHECK_NULL_PARAM(apAtomic, sub_atomic_exit);
 
 	if (apOldValue != NULL) {
-		*apOldValue = atomic_read(&apAtomic->mOSAtomic);
+		*apOldValue = (uint32_t)atomic_read(&apAtomic->mOSAtomic);
 	}
 
 	atomic_sub((int32_t)aSubValue, &apAtomic->mOSAtomic);
